Some more refactoring for onboarding
