+++ lib/constants.ts
@@ -5,5 +5,3 @@ export const TEAMNAMES = process.env.TEAM_NAMES;
 export const ORG = process.env.ORG;
 export const JOIN_PAGE = process.env.NEXT_PUBLIC_JOINING_PAGE_URL || "https://github.com/Dijkstra-Edu";
 export const DOMAIN = process.env.DOMAIN;
-
-
